+++ src/connectProvider/web3Provider/web3-provider/methods/eth_accounts.js
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 import { toPayload } from '../jsonrpc';
 
 export default async ({ payload, store }, res, next) => {