+++ src/connectProvider/web3Provider/web3-provider/methods/eth_sendTransaction.js
@@ -26,6 +26,7 @@ export default async (
   res,
   next
 ) => {
+  console.log(payload); // todo remove dev item
   if (payload.method !== 'eth_sendTransaction') return next();
   const tx = Object.assign({}, payload.params[0]);
   const localTx = Object.assign({}, tx);
@@ -65,7 +66,7 @@ export default async (
   }
   getSanitizedTx(tx)
     .then(_tx => {
-      console.log('TX', _tx); // todo remove dev item
+      debug('TX', _tx); // todo remove dev item
       eventHub.emit(EventNames.SHOW_TX_CONFIRM_MODAL, _tx, _response => {
         if (_response.reject) {
           debug('USER DECLINED SIGN TRANSACTION & SEND');
@@ -87,6 +88,17 @@ export default async (
                 ),
                 timestamp: localStoredObj.timestamp
               };
+              if(store.noSubs){
+                const txHash = hash;
+                let interval = setInterval(() => {
+                  store.state.web3.eth.getTransactionReceipt(txHash).then(result => {
+                    if(result !== null){
+                      clearInterval(interval)
+                      _promiObj.emit('receipt', res)
+                    }
+                  });
+                }, 1000);
+              }
             }
             res(null, toPayload(payload.id, hash));
           })